changed ES mapping. creating tokenized suggestions
