+++ go.mod
@@ -8,7 +8,7 @@ require (
 	github.com/davecgh/go-spew v1.1.1
 	github.com/elastic/go-elasticsearch/v7 v7.4.1-0.20191002090452-e26e3a7b7f9c // indirect
 	github.com/joho/godotenv v1.3.0 // indirect
-	github.com/lruggieri/utils v1.1.0
+	github.com/lruggieri/utils v1.1.1
 	github.com/nmmh/magneturi v0.0.0-20180607142838-464c677b3fb5
 	github.com/olivere/elastic/v7 v7.0.8
 	github.com/otiai10/gosseract v2.2.1+incompatible
